Moved sqlite update scripts to own folder and namespace.
